+++ packages/ui-components/src/map-legend/MapLegend.tsx
@@ -1,22 +1,10 @@
 import React, { memo } from 'react'
 import cx from 'classnames'
-import {
-  LayerMetadataLegend,
-  LayerMetadataLegendBivariate,
-} from '@globalfishingwatch/layer-composer/dist/types'
 import Solid from './Solid'
 import ColorRamp from './ColorRamp'
 import Bivariate from './Bivariate'
 import styles from './MapLegend.module.css'
-
-// TODO unify this with use-map-legend
-export type LegendLayer = LayerMetadataLegend & {
-  color: string
-}
-
-export type LegendLayerBivariate = LayerMetadataLegendBivariate & {
-  color: string
-}
+import { LegendLayer, LegendLayerBivariate } from '.'
 
 interface MapLegendProps {
   className?: string
@@ -33,12 +21,12 @@ interface MapLegendsProps {
 export function MapLegend({ layer, className, currentValueClassName }: MapLegendProps) {
   // TODO: include user context and categorical options
   if (layer.type === 'solid') {
-    return <Solid layer={layer} className={className} />
+    return <Solid layer={layer as LegendLayer} className={className} />
   }
   if (layer.type === 'colorramp' || layer.type === 'colorramp-discrete') {
     return (
       <ColorRamp
-        layer={layer}
+        layer={layer as LegendLayer}
         className={className}
         currentValueClassName={currentValueClassName}
       />
@@ -52,7 +40,6 @@ export function MapLegend({ layer, className, currentValueClassName }: MapLegend
 
 function MapLegends(props: MapLegendsProps) {
   const { className, layers } = props
-  console.log(layers)
   return (
     <div className={cx(styles.legend, className)}>
       {layers?.map((layer, index) => (
